enhance: 抽取位移动画逻辑，并修复点设置变成继续游戏的问题
